+++ src/amqp/src/Message/ConsumerMessageInterface.php
@@ -25,4 +25,6 @@ interface ConsumerMessageInterface extends MessageInterface
     public function isRequeue(): bool;
 
     public function getQueueBuilder(): QueueBuilder;
+
+    public function getConsumerTag(): string;
 }