+++ src/db-connection/src/Connection.php
@@ -116,7 +116,7 @@ class Connection extends BaseConnection implements ConnectionInterface, DbConnec
     public function release(): void
     {
         if ($this->isTransaction()) {
-            $this->rollBack();
+            $this->rollBack(0);
             $this->logger->error('Maybe you\'ve forgotten to commit or rollback the MySQL transaction.');
         }
         parent::release();